Replace homemade state system with Control.ST A single state machine (CLIApp) now handles gameplay and setting up games. This approach is much better and I anticipated expanding the game will now be significantly easier.
My goal was for the functionality to change as little as possible,
but there are some changes (both new features and regressions):
- You can enter 'exit' in the menu to cleanly end the program.
- When a game ends, you return to the menu instead of the program
closing.
- The number of wins for each player (and draws) are now stored for
the duration of a program's running time.
- The entire gamestate is displayed again after the user makes an
invalid command. This shouldn't happen, but the obvious fix would
remove totality from the IO interface implelemenation.
- Players are referred to only as 'P1' and 'P2', instead of more
descriptive names such as 'X' and 'Chaos.'